+++ packages/block-library/src/quote/index.js
@@ -19,13 +19,17 @@ import { G, Path, SVG } from '@wordpress/components';
 
 const blockAttributes = {
 	value: {
+		type: 'string',
 		source: 'html',
 		selector: 'blockquote',
 		multiline: 'p',
+		default: '',
 	},
 	citation: {
+		type: 'string',
 		source: 'html',
 		selector: 'cite',
+		default: '',
 	},
 	align: {
 		type: 'string',
@@ -36,7 +40,7 @@ export const name = 'core/quote';
 
 export const settings = {
 	title: __( 'Quote' ),
-	description: __( 'Maybe someone else said it better -- add some quoted text.' ),
+	description: __( 'Give quoted text visual emphasis. "In quoting others, we cite ourselves." — Julio Cortázar' ),
 	icon: <SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><Path fill="none" d="M0 0h24v24H0V0z" /><G><Path d="M19 18h-6l2-4h-2V6h8v7l-2 5zm-2-2l2-3V8h-4v4h4l-2 4zm-8 2H3l2-4H3V6h8v7l-2 5zm-2-2l2-3V8H5v4h4l-2 4z" /></G></SVG>,
 	category: 'common',
 	keywords: [ __( 'blockquote' ) ],
@@ -56,9 +60,12 @@ export const settings = {
 				blocks: [ 'core/paragraph' ],
 				transform: ( attributes ) => {
 					return createBlock( 'core/quote', {
-						value: toHTMLString( join( attributes.map( ( { content } ) =>
-							create( { html: content } )
-						), '\u2028' ), 'p' ),
+						value: toHTMLString( {
+							value: join( attributes.map( ( { content } ) =>
+								create( { html: content } )
+							), '\u2028' ),
+							multilineTag: 'p',
+						} ),
 					} );
 				},
 			},
@@ -113,7 +120,7 @@ export const settings = {
 							...split( create( { html: value, multilineTag: 'p' } ), '\u2028' )
 								.map( ( piece ) =>
 									createBlock( 'core/paragraph', {
-										content: toHTMLString( piece ),
+										content: toHTMLString( { value: piece } ),
 									} )
 								)
 						);
@@ -153,12 +160,15 @@ export const settings = {
 
 					return [
 						createBlock( 'core/heading', {
-							content: toHTMLString( pieces[ 0 ] ),
+							content: toHTMLString( { value: pieces[ 0 ] } ),
 						} ),
 						createBlock( 'core/quote', {
 							...attrs,
 							citation,
-							value: toHTMLString( quotePieces.length ? join( pieces.slice( 1 ), '\u2028' ) : create(), 'p' ),
+							value: toHTMLString( {
+								value: quotePieces.length ? join( pieces.slice( 1 ), '\u2028' ) : create(),
+								multilineTag: 'p',
+							} ),
 						} ),
 					];
 				},
@@ -292,8 +302,10 @@ export const settings = {
 			attributes: {
 				...blockAttributes,
 				citation: {
+					type: 'string',
 					source: 'html',
 					selector: 'footer',
+					default: '',
 				},
 				style: {
 					type: 'number',