+++ packages/compose/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.1.0 (Unreleased)
+## 2.1.0 (2018-10-29)
 
 ### Deprecation