+++ packages/notices/CHANGELOG.md
@@ -1,3 +1,3 @@
-## 1.0.0 (Unreleased)
+## 1.0.0 (2018-10-29)
 
 - Initial release.